feat: rename `deref` `useAtom`; add non-hook `deref`; `select` option
Rename `deref` to `useAtom` to conform with [the convention prescribed in the React
docs](https://github.com/reactjs/reactjs.org/blob/13959877d32cdac550c41f62f84e4866f2dc8f47/content/docs/hooks-custom.md#extracting-a-custom-hook)
and to benefit from the hooks eslint plugin. Implement a new version of `deref` that is not a React
custom hook for use outside of components. Enhance `useAtom` and `deref` with an options argument
that takes a `select` function to select or derive a value from atom state.

BREAKING CHANGE: `deref` is no longer a custom React hook and therefore function components that
call it will no longer be subscribed to the Atom's state. `deref` now only reads Atom state and is
therefore useful for reading Atom state from outside function components. The previous behavior of
`deref` is now available in the `useAtom` function, which is a custom React hook.