feat: add `initialize` function to initialize with alt hooks
This commit introduces a new way to initialize the library so that it is compatible with other
versions/implementations of hooks. My use case for this at the moment is to be able to use
pony/polyfills of hooks in react and react-native so that I can use hooks and react-atom in versions
of react prior to the hooks release.