feat: add @libre/atom dependency
This commit represents a refactoring of `dbeining/react-atom` where the implementations of Atom,
deref, set, and swap were extracted to an external dependency published in the registry under the
name `@libre/atom`. `@libre/atom` also introduces some API changes for `react-atom`, two by way of
addition and on by subtraction. `@libre/atom` adds the functions `addChangeHandler` and
`removeChangeHandler`, which allow for registering callbacks on Atoms to run whenever state changes.
`@libre/atom` also removes the function `atom`, which was simply an alias of `Atom.of`.

BREAKING CHANGE: remove `atom` function, which was an alias of `Atom.of`